requirements.txt
openai>=1.0.0
python-docx>=0.8.11
pandas>=1.5.0
numpy>=1.20.0
PyYAML>=5.4
progressbar2>=3.53.3